+++ examples/javaScript/subscriptions/listSubscriptions.js
@@ -1,0 +1,24 @@
+'use strict';
+
+var accessKey = process.env.AWS_ACCESS_KEY_ID || 'YOUR_KEY';
+var accessSecret = process.env.AWS_SECRET_ACCESS_KEY || 'YOUR_SECRET';
+
+var amazonMws = require('../../../lib/amazon-mws')(accessKey, accessSecret);
+
+var subscriptionRequest = function () {
+    amazonMws.subscriptions.searchFor({
+        'Version': '2013-07-01',
+        'Action': 'ListSubscriptions',
+        'SellerId': 'SELLER_ID',
+        'MWSAuthToken': 'MWS_AUTH_TOKEN',
+        'MarketplaceId': 'MARKET_PLACE_ID'
+    }, function (error, response) {
+        if (error) {
+            console.log('error ', error);
+            return;
+        }
+        console.log('response', response);
+    });
+};
+
+subscriptionRequest();